node: Use new routing datapath for node events
This commit achieves separation of node definition, node management and
datapath implementation based on node events.

pkg/node contains the definition of a node, its addressing, logic to determine
the primary external address, determination of local hostname, and management
of the addressing for the local node. The addressing should be moved out of
this package in a future PR but was outside of thes scope of this commit.

pkg/node/manager manages the list of nodes and implements the logic to handle
node events from multiple sources. Sources include:

    * local node definition
    * local agent information
    * kubernetes
    * kvstore

Two new metrics are introduced:

   * gauge cilium_nodes_all_num
   * counter cilium_nodes_all_events_received_total {"all", "update", "delete"}
   * counter cilium_nodes_all_datapath_validations_total

This allows to visualize the current cluster size as well as the rate of node
updates received per minute.

The manager periodically requests the datapath to validate the implementation
of a node. As the cost of validation grows with the cluster size, the interval
in which the validation is performed is made dependent on the cluster size
using a logarithmic function.

The local node management is moved into daemon/node_discovery.go and simplified
by allocating health IPs before starting node discovery. This avoids requiring
to annodate the k8s node multiple times.

Signed-off-by: Thomas Graf <thomas@cilium.io>